+++ tests/framework/grid/CheckboxColumnTest.php
@@ -54,6 +54,35 @@ class CheckboxColumnTest extends TestCase
         $this->assertContains('name="MyForm[grid1][key_all]"', $column->renderHeaderCell());
     }
 
+    public function testInputValue()
+    {
+        $column = new CheckboxColumn();
+        $this->assertContains('value="1"', $column->renderDataCell([], 1, 0));
+        $this->assertContains('value="42"', $column->renderDataCell([], 42, 0));
+        $this->assertContains('value="[1,42]"', $column->renderDataCell([], [1, 42], 0));
+
+        $column = new CheckboxColumn(['checkboxOptions' => ['value' => 42]]);
+        $this->assertNotContains('value="1"', $column->renderDataCell([], 1, 0));
+        $this->assertContains('value="42"', $column->renderDataCell([], 1, 0));
+
+        $column = new CheckboxColumn([
+            'checkboxOptions' => function ($model, $key, $index, $column) {
+                return [];
+            }
+        ]);
+        $this->assertContains('value="1"', $column->renderDataCell([], 1, 0));
+        $this->assertContains('value="42"', $column->renderDataCell([], 42, 0));
+        $this->assertContains('value="[1,42]"', $column->renderDataCell([], [1, 42], 0));
+
+        $column = new CheckboxColumn([
+            'checkboxOptions' => function ($model, $key, $index, $column) {
+                return ['value' => 42];
+            }
+        ]);
+        $this->assertNotContains('value="1"', $column->renderDataCell([], 1, 0));
+        $this->assertContains('value="42"', $column->renderDataCell([], 1, 0));
+    }
+
     /**
      * @return GridView a mock gridview
      */